project/social_media_sentiment/requirements.txt
# Core ML and NLP
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99

# Financial data
yfinance>=0.2.28
requests>=2.31.0

# Social Media APIs - Tier 1
praw>=7.7.1  # Reddit API
tweepy>=4.14.0  # Twitter/X API
beautifulsoup4>=4.12.0  # Web scraping (StockTwits, Telegram)
lxml>=4.9.0  # HTML parser

# Social Media APIs - Tier 2
google-api-python-client>=2.100.0  # YouTube Data API
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Text processing & filtering
nltk>=3.8.0
textblob>=0.17.1
langdetect>=1.0.9

# Web scraping utilities
selenium>=4.15.0  # Optional: for dynamic content
html5lib>=1.1

# Apify (live data collection)
apify-client>=1.6.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
dateparser>=1.1.8

# Dataset loading (HuggingFace)
datasets>=2.14.0
pyarrow>=12.0.0
